+++ project/frontend/src/components/PreferenceForm.css
@@ -1,18 +1,26 @@
-.preference-form-container {
-  max-width: 600px;
+.form-container {
+  /* max-width: 600px; */
+  max-width: 800px;
   text-align: left;
+  margin: 0px auto;
 }
 
-.preference-form-container button {
+.form-container button {
   display: block;
   margin: 16px auto;
 }
 
-.preference-form {
+.form {
   margin-bottom: 0px;
 }
 
-.preference-form-submit-container {
+.header {
+  text-align: center;
+  margin-bottom: 20px;
+  margin-top: 15px;
+}
+
+.form-submit-container {
   position: sticky;
   bottom: 0px;
   background-color: white;
@@ -20,83 +28,6 @@
   border-top: 1px solid #bbb;
 }
 
-.preference-form-submit-container button {
+.form-submit-container button {
   margin: 0px auto;
-}
-
-.preference-form-column {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.preference-form-column > .preference-form-row:first-child {
-  margin: 0px;
-}
-
-.preference-form-column > .preference-form-row:last-child {
-  margin: 0px;
-}
-
-.preference-form-row {
-  display: flex;
-  align-items: center;
-  margin: 16px 0px;
-}
-
-.preference-form-row label {
-  margin: 0px 16px;
-  white-space: nowrap;
-}
-
-.preference-form-row p {
-  margin-bottom: 0px;
-  padding: 6px 8px;
-  border: 1px solid #bbb;
-  border-radius: 4px;
-  background-color: #f3f3f3;
-  font-size: 1.4rem;
-}
-
-.preference-form-row select {
-  width: 100%;
-  margin: 0px;
-  font-family: 'Roboto';
-  font-size: 1.4rem;
-}
-
-.preference-form-row input {
-  margin: 0px;
-}
-
-.preference-form-divider {
-  height: 64px;
-  display: flex;
-  align-items: center;
-}
-
-.preference-form-divider p {
-  color: #555;
-  text-align: center;
-  font-size: 11px;
-  font-weight: 600;
-  line-height: 38px;
-  letter-spacing: 0.1rem;
-  text-transform: uppercase;
-  text-decoration: none;
-  white-space: nowrap;
-  font-family: 'Montserrat';
-  margin: 0px 8px 0px 8px;
-}
-
-.preference-form-divider div {
-  display: inline-block;
-  width: 100%;
-  height: 0.5px;
-  background-color: #bbb;
-}
-
-.preference-form-slider-container {
-  width: 100%;
-  padding: 16px 0px;
 }
